refactor: Move `ActivityChip` to `components` folder
I was waiting for kantahrek's PR to go up to avoid creating
undue merge conflicts.

Now that we're both at the same version, the refactor can be
completed safely.
